pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdfetch"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pypdf>=4.0",
    "pymupdf>=1.24",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
pdfetch = "pdfetch.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["pdfetch*"]
